added 4 new functions to a new file called userstats
